Unsigned short promotes to signed int, so cast to unsigned instead.
git-svn-id: svn+ssh://trac.godzil.net/gcc4ti@527 9552661e-59e3-4036-b4f2-dbe53926924f

+++ trunk/ktigcc/preferences.cxx
@@ -1152,6 +1152,6 @@ void savePreferences(TIGCCPrefs *prefs,KConfig *cfg)
     cfg->setGroup("Preferences");
     cfg->writeEntry("Lazy Loading",(bool)prefs->lazyLoading);
     cfg->writeEntry("Use Calc Charset",(bool)prefs->useCalcCharset);
-    cfg->writeEntry("Tab Width Asm",(unsigned short)prefs->tabWidthAsm);
-    cfg->writeEntry("Tab Width C",(unsigned short)prefs->tabWidthC);
+    cfg->writeEntry("Tab Width Asm",(unsigned)prefs->tabWidthAsm);
+    cfg->writeEntry("Tab Width C",(unsigned)prefs->tabWidthC);
 }
